+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # Changelog for v0.2.2
-
+* improve shell and add commands for translate, save_as